Rename a few things for organization
